requirements.txt
fastapi
uvicorn
pydantic
numpy
faiss-cpu
sentence-transformers
rich
